requirements.txt
fastapi==0.112.0
pyodbc==5.1.0
python-dotenv==1.0.1
requests==2.32.3
sqlalchemy==2.0.32
uvicorn==0.30.5